post email with a custom hook
